Started to work on the client side and reworked the file system a bit

+++ src/ru/loghorrean/veganShop/models/database/entities/CustomDishInOrder.java
@@ -3,9 +3,9 @@ package ru.loghorrean.veganShop.models.database.entities;
 public class CustomDishInOrder extends DatabaseEntity {
     private CustomDish dish;
     private Order order;
-    private float amount;
+    private int amount;
 
-    public CustomDishInOrder(int id, CustomDish dish, Order order, float amount) {
+    public CustomDishInOrder(int id, CustomDish dish, Order order, int amount) {
         super(id);
         this.dish = dish;
         dish.addOrder(order);
@@ -14,7 +14,7 @@ public class CustomDishInOrder extends DatabaseEntity {
         this.amount = amount;
     }
 
-    public CustomDishInOrder(CustomDish dish, Order order, float amount) {
+    public CustomDishInOrder(CustomDish dish, Order order, int amount) {
         this(-1, dish, order, amount);
     }
 
@@ -38,11 +38,11 @@ public class CustomDishInOrder extends DatabaseEntity {
         this.order.addCustomDish(dish);
     }
 
-    public float getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
